requirements.txt
fastapi
uvicorn
requests
pymongo
python-dotenv
